refactor: MenuGroup JPA로 변경
