+++ src/main/java/kitchenpos/domain/repository/MenuGroupDao.java
@@ -1,0 +1,9 @@
+package kitchenpos.domain.repository;
+
+import kitchenpos.domain.MenuGroup;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MenuGroupDao extends JpaRepository<MenuGroup, Long> {
+
+    boolean existsById(Long id);
+}
