+++ src/main/java/kitchenpos/ui/MenuGroupController.java
@@ -1,37 +1,38 @@
 package kitchenpos.ui;
 
+import java.net.URI;
+import java.util.List;
 import kitchenpos.application.MenuGroupService;
 import kitchenpos.domain.MenuGroup;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.net.URI;
-import java.util.List;
-
 @RestController
-public class MenuGroupRestController {
+@RequestMapping("/api/menu-groups")
+public class MenuGroupController {
+
     private final MenuGroupService menuGroupService;
 
-    public MenuGroupRestController(final MenuGroupService menuGroupService) {
+    public MenuGroupController(final MenuGroupService menuGroupService) {
         this.menuGroupService = menuGroupService;
     }
 
-    @PostMapping("/api/menu-groups")
+    @PostMapping
     public ResponseEntity<MenuGroup> create(@RequestBody final MenuGroup menuGroup) {
         final MenuGroup created = menuGroupService.create(menuGroup);
         final URI uri = URI.create("/api/menu-groups/" + created.getId());
+
         return ResponseEntity.created(uri)
-                .body(created)
-                ;
+            .body(created);
     }
 
-    @GetMapping("/api/menu-groups")
+    @GetMapping
     public ResponseEntity<List<MenuGroup>> list() {
         return ResponseEntity.ok()
-                .body(menuGroupService.list())
-                ;
+            .body(menuGroupService.list());
     }
 }
